Refactor helpers to not interact with event, remove unneeded config
This commit refactors all the helper functions to not receive the event or modify the unit status. They are instead agnostic of the events and throw exceptions
when they fail. The callers (i.e. hook functions) are the ones catching the exceptions and updating the unit status accordingly.

Also, removes unneeded config `from_resources` (it can be inferred from the presence or absence of resources at install time).

Also makes some small improvements, such as comparing hashes of installed debs when upgrading, for installations from resources.